+++ packages/axios-extension/test/abap/lrep-service.test.ts
@@ -166,6 +166,51 @@ describe('LayeredRepositoryService', () => {
                 expect(error).toBeDefined();
             }
         });
+
+        test('try undeploying on a too old ABAP system', async () => {
+            nock(server)
+                .get((url) => {
+                    return url.startsWith(
+                        `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                            config.namespace as string
+                        )}&layer=CUSTOMER_BASE`
+                    );
+                })
+                .reply(200, undefined, {
+                    'x-csrf-token': 'token'
+                });
+            nock(server)
+                .delete(
+                    `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                        config.namespace as string
+                    )}&layer=CUSTOMER_BASE&changelist=${config.transport}`
+                )
+                .reply(405);
+            try {
+                await service.undeploy(config);
+                fail('The function should have thrown an error.');
+            } catch (error) {
+                expect(error).toBeDefined();
+            }
+        });
+    });
+
+    describe('getCsrfToken', () => {
+        test('successful call', async () => {
+            nock(server).get(`${LayeredRepositoryService.PATH}/actions/getcsrftoken/`).reply(200);
+            const response = await service.getCsrfToken();
+            expect(response).toBeDefined();
+        });
+
+        test('error is thrown', async () => {
+            nock(server).get(`${LayeredRepositoryService.PATH}/actions/getcsrftoken/`).reply(403);
+            try {
+                await service.getCsrfToken();
+                fail('The function should have thrown an error.');
+            } catch (error) {
+                expect(error).toBeDefined();
+            }
+        });
     });
 
     describe('mergeAppDescriptorVariant', () => {