+++ packages/deploy-tooling/src/base/deploy.ts
@@ -282,7 +282,13 @@ async function tryDeploy(
         if (isBspConfig(config.app) && !config.lrep) {
             if (config.test === true) {
                 const validateOutput = await validateBeforeDeploy(config, provider, logger);
-                logger.info(formatSummary(validateOutput.summary));
+                if (!validateOutput.result) {
+                    logger.info(
+                        `Results of validating the deployment configuration settings:${formatSummary(
+                            validateOutput.summary
+                        )}`
+                    );
+                }
             }
             const service = getDeployService(provider.getUi5AbapRepository, config, logger);
             await service.deploy({ archive, bsp: config.app, testMode: config.test, safeMode: config.safe });